PMFRG.jl.

Function Documentation

OrdinaryDiffEq.BS3 — Type

BS3(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
        step_limiter! = OrdinaryDiffEq.trivial_limiter!,
        thread = OrdinaryDiffEq.False(),)

Explicit Runge-Kutta Method. A third-order, four-stage explicit FSAL Runge-Kutta method with embedded error estimator of Bogacki and Shampine.

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False(), default) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

@article{bogacki19893, title={A 3 (2) pair of Runge-Kutta formulas}, author={Bogacki, Przemyslaw and Shampine, Lawrence F}, journal={Applied Mathematics Letters}, volume={2}, number={4}, pages={321–325}, year={1989}, publisher={Elsevier} }

PMFRG.BSE_iteration! — Function

Performs single BSE iteration. Specify State explicitly to make fixed point libraries compatible (even though it may point to the same memory location as Workspace.State)

PMFRG.BareVertexType — Type

Stores all information about the bare vertex (without frequencies)

PMFRG.BareVertex_Freq — Function

Allocates a frequency-dependent vertex type and initializes it with the bare vertex

PMFRG.BubbleType — Type

Struct storing information about Bubble functions, i.e. rhs derivatives of vertex flow equations.

OrdinaryDiffEq.DP5 — Type

DP5(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
        step_limiter! = OrdinaryDiffEq.trivial_limiter!,
        thread = OrdinaryDiffEq.False(),)

Explicit Runge-Kutta Method. Dormand-Prince's 5/4 Runge-Kutta method. (free 4th order interpolant).

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False(), default) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

@article{dormand1980family, title={A family of embedded Runge-Kutta formulae}, author={Dormand, John R and Prince, Peter J}, journal={Journal of computational and applied mathematics}, volume={6}, number={1}, pages={19–26}, year={1980}, publisher={Elsevier} }

PMFRG.Lam_to_t — Function

Converts physical cutoff Λ to t (integrator step). Inverse of ttoLam.

PMFRG.MultiLoopWorkSpace — Type

Struct containing all memory used in a single ODE step

PMFRG.NumericalParams — Type

Struct to hold all relevant quantities that are needed throughout the computation.

N::Int = 24
Ngamma::Int = 100 #Number of gamma frequencies
VDims::NTuple{4,Int} = (Npairs,N,N,N)
accuracy::Float64 = 1e-6
Lam_min::Float64 = exp(-10.),
Lam_max::Float64 = exp(10.),
usesymmetry::Bool = true
MinimalOutput::Bool = false
ex_freq::Float64 = (2*N-1)*pi*T
lenIntw::Int = N
lenIntw_acc::Int = 2*maximum((N,Ngamma,lenIntw)) # more accurate for less demanding sums
np_vec::Array{Int,1} = collect(0:N-1)
np_vec_gamma::Array{Int,1} = collect(0:Ngamma-1)

PMFRG.Observables — Type

Struct containing the observables that are saved at every step

PMFRG.OneLoopWorkspace — Type

Convenience struct containing references to arrays for vertices and their derivative. This does not allocate any memory and is performant.

PMFRG.PMFRGParams — Type

Abstract Param struct to dispatch between different PMFRG methods (i.e. Two-Loop or Parquet) Assumed to have at least the fields: System::Geometry NumericalParams::NumericalParams Options

PMFRG.Params — Function

Construct a set of parameters used for an FRG calculation. 'System' refers to the geometry that is used for the calculation (see SpinFRGLattices). The second argument is the method, i.e. OneLoop() (default value), TwoLoop(), Parquet(). For possible optional keyword arguments see the docs of 'NumericalParams'.

Params constructor for parquet calculation. For all available params, see also PMFRG.ParquetOptions and PMFRG.NumericalParams.

PMFRG.SetCompletionCheckmark — Function

Rename CurrentState to FinalState as indicator that Job is finished

PMFRG.SetupParquet — Function

Constructs Workspace for parquet equations. Irreducible vertex can later be provided by optional argument but this is currently not supported

PMFRG.SolveFRG — Function

Given a set of Parameters (currently, oneß and twoloop are implemented), solves the set of differential flow equations and returns the ODE solution object along with an array of Observables at each lambda step. Allowed keyword arguments (with default values):

MainFile = nothing,                             # Specifies name of main output file as a string.
                                                # Defaults to 'nothing', in which case no output file is produced.
Group = DefaultGroup(Par),               # Specifies the name of the subgroup of the main files HDF5 group to which the output data is written. Defaults to temperature
                                                # Defaults to the value of temperature to allow temperature sweeps to be written to the same file.
CheckpointDirectory = nothing,                  # Directory in which the current ODE state is written in regular intervals during the integration. 
                                                # Default 'nothing' will not produce any backup!
method = DP5(),                                 # ODE integration method. Standard is set to OrdinaryDiffEq.DP5(). 
                                                # See OrdinaryDiffEq package for further options.
MaxVal = Inf                                    # Terminates the ODE solution, when any absolute value of the Solution reaches MaxVal.
ObsSaveat = nothing,                            # Specifies a vector of Λ values at which Observables are computed.
VertexCheckpoints = [],                         # Specifies a vector of Λ values at which vertices shall be saved permanently without being overwritten. 
                                                # Empty means only the current state will be saved. Requires CheckpointDir ≠ nothing !
overwrite_Checkpoints = false::Bool,            # Specifies whether CheckpointDirectory is to be overwritten if it exists. 
                                                # Defaults to false, in which case an unused name is generated.
CheckPointSteps = 1,                            # Number of skipped steps before Checkpoint data is saved to reduce time spent on IO operations.
kwargs...                                       # Additional keyword arguments are passed to OrdinaryDiffEq.solve.
                                                # See the OrdinaryDiffEq documentation for further details.

Solves FRG as specified for parameters

PMFRG.StateType — Type

Struct containing information about the (physical) ODE State, i.e. vertices

PMFRG.TwoLoopWorkspace — Type

Convenience struct containing references to arrays for vertices and their derivative. This does not allocate any memory and is performant.

PMFRG.V_ — Function

Returns value of vertex, swaps sites i <-> j by reading from inverted pairs when necessary. For performance, it is advised to directly use the Arrays like Va after pre-checking the frequency structure

OrdinaryDiffEq.Vern7 — Type

Vern7(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
        step_limiter! = OrdinaryDiffEq.trivial_limiter!,
        thread = OrdinaryDiffEq.False(),
        lazy = true)

Explicit Runge-Kutta Method. Verner's “Most Efficient” 7/6 Runge-Kutta method. (lazy 7th order interpolant).

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False(), default) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

lazy: determines if the lazy interpolant is used.

References

@article{verner2010numerically, title={Numerically optimal Runge–Kutta pairs with interpolants}, author={Verner, James H}, journal={Numerical Algorithms}, volume={53}, number={2-3}, pages={383–396}, year={2010}, publisher={Springer} }

PMFRG.VertexType — Type

Struct containing information about four-point vertices

PMFRG._convertToFloat — Function

Converts other arguments to type of first argument. If that is not a float, converts everything to Float64.

PMFRG.addBR! — Function

adds to ResultBubble given the vertex as well as a bubble inserted on the left. Assumes that vertices and bubbles are given pre-computed in VertexBuffer.

PMFRG.addTo1PartBubble! — Function

Computes a single-particle (i.e. self-energy) bubble. Allows specification of function type, i.e. what vertices are used since this is different if a bubble function is inserted as opposed to a vertex.

Computes a single-particle (i.e. self-energy) bubble. Can only be used if B is a bubble function

Computes a single-particle (i.e. self-energy) bubble. Can only be used if argument is a vertex

PMFRG.addTo2PartBubble! — Function

Computes a two-particle bubble in the s-Channel given two four-point functions (i.e. vertices or other bubbles). Γ is assumed to be a vertex. To allow the computation using just the left part of a bubble, specification of the transpose is needed i.e. Transpose(XL) = XR and Transpose(X) = X, where X = XL + XR is the full bubble.

PMFRG.addX! — Function

adds part of X functions in Matsubara sum at nwpr containing the site summation for a set of s t and u frequencies. This is the most numerically demanding part!

Use multiple dispatch to treat the common special case in which the propagator does not depend on site indices to increase performance

PMFRG.addXTilde! — Function

Use multiple dispatch to treat the common special case in which the propagator does not depend on site indices to increase performance

PMFRG.gamma_ — Function

Returns self energy evaluated at the appropriate frequency

PMFRG.getXBubblePartition! — Function

writing to X and XTilde in Workspace, computes bubble diagrams within a range of frequencies given by isrange, itrange and iurange

PMFRG.get_sign_iw — Function

Given a (Matsubara) integer, returns a corresponding signed integer within the considered range

PMFRG.get_w — Function

given a Matsubara (!) integer, return the corresponding Matsubara frequency

PMFRG.iG_ — Function

Taking a Matsubara integer, gives fully dressed propagator

PMFRG.iSKat_ — Function

Taking a Matsubara integer, gives full single-scale propagator with Katanin (vertex flow equation)

PMFRG.iS_ — Function

Taking a Matsubara integer, gives single-scale propagator without Katanin (self energy flow equation)

PMFRG.iterateSDE! — Function

Self-consistently iterates SDE until convergence is reached.

PMFRG.iterateSolution! — Function

Obtains a solution to Bethe-Salpeter and Schwinger-Dyson equations by iteration until convergence is reached up to accuracy specified by accuracy in Params

PMFRG.joinGroup — Function

This file contains functions used for reading and writing FRG data to disc. Stored data can be used to re-launch an incomplete FRG calculation.

PMFRG.readState — Function

Reads Vertices from file

PMFRG.saveGeometryParams — Function

Saves important information about Geometry parameters so that they can be reconstructed

PMFRG.saveNumericalParams — Function

Saves important information about computation parameters so that they can be reconstructed

PMFRG.saveObs — Function

Saves Observables

PMFRG.saveParams — Function

Saves important information about parameters so that they can be reconstructed

PMFRG.saveState — Function

Saves Vertices to a compressed HDF5 file in a Group "Lam".

PMFRG.setZero! — Function

fills an array with zeros

Recursively sets structure to zero

PMFRG.squareDist — Function

Computes total difference between two objects of type T by summing up the distance of all elements

PMFRG.symmetrizeBubble! — Function

Use symmetries and identities to compute the rest of bubble functions

PMFRG.t_to_Lam — Function

Converts t step used for integrator to Λ. Inverse of Lamtot.

PMFRG.writeTo! — Function

Set two objects equal by recusively setting all their fields to be equal

PMFRG.SolveParquet — Function

Given set of parameters solve the self-consistent Parquet approximation iteratively. TODO: Save output to file