PMFRG.jl.

Function Documentation

PMFRG.SolveFRG — Function

Given a set of Parameters (currently, oneß and twoloop are implemented), solves the set of differential flow equations and returns the ODE solution object along with an array of Observables at each lambda step. Allowed keyword arguments (with default values):

MainFile = nothing,                             # Specifies name of main output file as a string.
                                                # Defaults to 'nothing', in which case no output file is produced.
Group = DefaultGroup(Par),               # Specifies the name of the subgroup of the main files HDF5 group to which the output data is written. Defaults to temperature
                                                # Defaults to the value of temperature to allow temperature sweeps to be written to the same file.
CheckpointDirectory = nothing,                  # Directory in which the current ODE state is written in regular intervals during the integration. 
                                                # Default 'nothing' will not produce any backup!
method = DP5(),                                 # ODE integration method. Standard is set to OrdinaryDiffEq.DP5(). 
                                                # See OrdinaryDiffEq package for further options.
MaxVal = Inf                                    # Terminates the ODE solution, when any absolute value of the Solution reaches MaxVal.
ObsSaveat = nothing,                            # Specifies a vector of Λ values at which Observables are computed.
VertexCheckpoints = [],                         # Specifies a vector of Λ values at which vertices shall be saved permanently without being overwritten. 
                                                # Empty means only the current state will be saved. Requires CheckpointDir ≠ nothing !
overwrite_Checkpoints = false::Bool,            # Specifies whether CheckpointDirectory is to be overwritten if it exists. 
                                                # Defaults to false, in which case an unused name is generated.
CheckPointSteps = 1,                            # Number of skipped steps before Checkpoint data is saved to reduce time spent on IO operations.
kwargs...                                       # Additional keyword arguments are passed to OrdinaryDiffEq.solve.
                                                # See the OrdinaryDiffEq documentation for further details.

Solves FRG as specified for parameters

PMFRG.Params — Function

Construct a set of parameters used for an FRG calculation. 'System' refers to the geometry that is used for the calculation (see SpinFRGLattices). The second argument is the method, i.e. OneLoop() (default value), TwoLoop(), Parquet(). For possible optional keyword arguments see the docs of 'NumericalParams'.

Params constructor for parquet calculation. For all available params, see also PMFRG.ParquetOptions and PMFRG.NumericalParams.

Missing docstring.

Missing docstring for OneLoopParams. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BS3. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Vern7. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DP5. Check Documenter's build log for details.

Missing docstring.

Missing docstring for version. Check Documenter's build log for details.

Missing docstring.

Missing docstring for getChi. Check Documenter's build log for details.

Missing docstring.

Missing docstring for OneLoop. Check Documenter's build log for details.

PMFRG.saveState — Function

Saves Vertices to a compressed HDF5 file in a Group "Lam".

PMFRG.readState — Function

Reads Vertices from file

Missing docstring.

Missing docstring for readLam. Check Documenter's build log for details.

PMFRG.saveParams — Function

Saves important information about parameters so that they can be reconstructed

Missing docstring.

Missing docstring for readParams. Check Documenter's build log for details.

Missing docstring.

Missing docstring for setupDirectory. Check Documenter's build log for details.

Missing docstring.

Missing docstring for saveCurrentState. Check Documenter's build log for details.

Missing docstring.

Missing docstring for UniqueDirName. Check Documenter's build log for details.

Missing docstring.

Missing docstring for UniqueFileName. Check Documenter's build log for details.

Missing docstring.

Missing docstring for generateName. Check Documenter's build log for details.

Missing docstring.

Missing docstring for setupFromCheckpoint. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SolveFRG_Checkpoint. Check Documenter's build log for details.

Missing docstring.

Missing docstring for readGeometry. Check Documenter's build log for details.

Missing docstring.

Missing docstring for readObservables. Check Documenter's build log for details.

Missing docstring.

Missing docstring for getUnfinishedJobs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for generateFileName. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TwoLoop. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MultiLoop. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Parquet. Check Documenter's build log for details.

PMFRG.SolveParquet — Function

Given set of parameters solve the self-consistent Parquet approximation iteratively. TODO: Save output to file